Phase1

+++ Scripts/Expansion/Original UO/Mobiles/Monsters/Gazers.cs
@@ -48,20 +48,9 @@ namespace Server.Mobiles
         {
         }
 
-        public override int TreasureMapLevel
-        {
-            get
-            {
-                return 1;
-            }
-        }
-        public override int Meat
-        {
-            get
-            {
-                return 1;
-            }
-        }
+        public override int TreasureMapLevel { get { return 1; } }
+        public override int Meat { get { return 1; } }
+
         public override void GenerateLoot()
         {
             AddLoot(LootPack.Average);
@@ -127,13 +116,7 @@ namespace Server.Mobiles
         {
         }
 
-        public override int TreasureMapLevel
-        {
-            get
-            {
-                return Core.AOS ? 4 : 0;
-            }
-        }
+        public override int TreasureMapLevel { get { return Core.AOS ? 4 : 0; } }
 
         public override void GenerateLoot()
         {
@@ -143,7 +126,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
