+++ Scripts/Expansion/Original UO/Mobiles/Monsters/SilverSerpent.cs
@@ -1,4 +1,3 @@
-using System;
 using Server.Factions;
 using Server.Items;
 
@@ -75,7 +74,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
