+++ Scripts/Expansion/Original UO/Mobiles/Monsters/Slime.cs
@@ -1,4 +1,3 @@
-using System;
 
 namespace Server.Mobiles
 {
@@ -89,7 +88,7 @@ namespace Server.Mobiles
         public override void Serialize(GenericWriter writer)
         {
             base.Serialize(writer);
-            writer.Write((int)1);
+            writer.Write(1);
         }
 
         public override void Deserialize(GenericReader reader)
